requirements.txt
requests
discord.py>=1.3.4 
riotwatcher
playwright
python-dotenv